Defined undefined behavior
Instructions between hardware does NOT have parity. For example, on arm64, INT_MIN / -1 = INT_MIN while on x86 the cpu throws an overflow exception. Also, shifting is different when the shift is greater than the bit width. But adhering to each architecture's definition on how to deal with fringe cases creates the need for extra unneeded instructions. So, instructions with non parity can optionally be checked within the pipeline.